requirements.txt
convertapi==1.8.0
cvzone==1.6.1
numpy==1.21.5
opencv-python==4.9.0
pdf2image==1.17.0
python-dotenv==1.0.1
